+++ src/autlib3.c
@@ -135,8 +135,8 @@ stpnlp(iap_type *iap, rap_type *rap, double *par, int64 *icp, double *u) {
 
     int64 ndm, ips, irs;
 
-    f = malloc(sizeof(*f)*(iap->ndim));
-    v = malloc(sizeof(*v)*(iap->ndim));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    v = malloc(sizeof(*v)*(usize)(iap->ndim));
     /* Generates starting data for the continuation of folds. */
 
     /* Local */
@@ -184,7 +184,7 @@ fnc1(iap_type *iap, rap_type *rap, int64 ndim, double *u, double *uold,
     double ddp[NPARX], *ddu;
     int64 ndm;
 
-    ddu = malloc(sizeof(*ddu)*(iap->ndim));
+    ddu = malloc(sizeof(*ddu)*(usize)(iap->ndim));
     /* Generate the equations for the continuation scheme used for */
     /* the optimization of algebraic systems (one parameter). */
 
@@ -337,7 +337,7 @@ ffc2(iap_type *iap, rap_type *rap, int64 ndim, double *u, double *uold,
     double ddp[NPARX], *ddu, fop;
     int64 ndm2;
 
-    ddu = malloc(sizeof(*ddu)*(iap->ndim));
+    ddu = malloc(sizeof(*ddu)*(usize)(iap->ndim));
     /* Local */
 
     /* Parameter adjustments */
@@ -399,10 +399,10 @@ stpnc2(iap_type *iap, rap_type *rap, double *par, int64 *icp, double *u) {
     double fop;
     int64 irs;
 
-    f = malloc(sizeof(*f)*(iap->ndim));
-    v = malloc(sizeof(*v)*(iap->ndim));
-    dd = malloc(sizeof(*dd)*(iap->ndim)*(iap->ndim));
-    du = malloc(sizeof(*du)*(iap->ndim));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    v = malloc(sizeof(*v)*(usize)(iap->ndim));
+    dd = malloc(sizeof(*dd)*(usize)((iap->ndim)*(iap->ndim)));
+    du = malloc(sizeof(*du)*(usize)(iap->ndim));
     /* Generates starting data for the continuation equations for */
     /* optimization of algebraic systems (More than one parameter). */
 
@@ -683,9 +683,9 @@ stpnhd(iap_type *iap, rap_type *rap, double *par, int64 *icp, double *u) {
 
     int64 ndm, irs, ndm2;
 
-    f = malloc(sizeof(*f)*(iap->ndim));
-    v = malloc(sizeof(*v)*(iap->ndim));
-    smat = malloc(sizeof(*smat)*(iap->ndim*2)*(iap->ndim*2));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    v = malloc(sizeof(*v)*(usize)(iap->ndim));
+    smat = malloc(sizeof(*smat)*(usize)((iap->ndim*2)*(iap->ndim*2)));
     /* Generates starting data for the continuation of Hopf bifurcation */
     /* points for maps. */
 
@@ -892,9 +892,9 @@ stpnhb(iap_type *iap, rap_type *rap, double *par, int64 *icp, double *u) {
     int64 ndm, irs;
     double rom;
     int64 ndm2;
-    smat = malloc(sizeof(*smat)*(iap->ndim*2)*(iap->ndim*2));
-    f = malloc(sizeof(*f)*(iap->ndim));
-    v = malloc(sizeof(*v)*(iap->ndim));
+    smat = malloc(sizeof(*smat)*(usize)((iap->ndim*2)*(iap->ndim*2)));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    v = malloc(sizeof(*v)*(usize)(iap->ndim));
     /* Generates starting data for the 2-parameter continuation of */
     /* Hopf bifurcation point (ODE). */
 
@@ -1100,11 +1100,11 @@ stpnhw(iap_type *iap, rap_type *rap, double *par, int64 *icp, double *u) {
     double rom;
     int64 ndm2;
 
-    smat = malloc(sizeof(*smat)*(2*iap->ndim)*(2*iap->ndim));
-    f = malloc(sizeof(*f)*(iap->ndim));
-    v = malloc(sizeof(*v)*(iap->ndim));
-    dfp = malloc(sizeof(*dfp)*(iap->ndim)*NPARX);
-    dfu = malloc(sizeof(*dfu)*(iap->ndim)*(iap->ndim));
+    smat = malloc(sizeof(*smat)*(usize)((2*iap->ndim)*(2*iap->ndim)));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    v = malloc(sizeof(*v)*(usize)(iap->ndim));
+    dfp = malloc(sizeof(*dfp)*(usize)((iap->ndim)*NPARX));
+    dfu = malloc(sizeof(*dfu)*(usize)((iap->ndim)*(iap->ndim)));
 
     /* Generates starting data for the continuation of a bifurcation to a */
     /* traveling wave. */
@@ -1396,10 +1396,10 @@ stpnps(iap_type *iap, rap_type *rap, double *par, int64 *icp, int64 *ntsr,
     double tpi;
     int64 irs;
 
-    smat = malloc(sizeof(*smat)*(iap->ndim*2)*(iap->ndim*2));
-    rnllv = malloc(sizeof(*rnllv)*(iap->ndim*2)*(iap->ndim*2));
-    f = malloc(sizeof(*f)*(iap->ndim));
-    u = malloc(sizeof(*u)*(iap->ndim));
+    smat = malloc(sizeof(*smat)*(usize)((iap->ndim*2)*(iap->ndim*2)));
+    rnllv = malloc(sizeof(*rnllv)*(usize)((iap->ndim*2)*(iap->ndim*2)));
+    f = malloc(sizeof(*f)*(usize)(iap->ndim));
+    u = malloc(sizeof(*u)*(usize)(iap->ndim));
     /* Generates starting data for the continuation of a branch of periodic */
     /* solutions from a Hopf bifurcation point. */
